requirements.txt
absl-py==0.9.0
astor==0.8.1
bleach==1.5.0
gast==0.3.3
grpcio==1.27.2
html5lib==0.9999999
Markdown==3.2.1
numpy==1.14.2
opencv-python==4.2.0.32
protobuf==3.11.3
six==1.14.0
tensorboard==1.6.0
tensorflow-gpu==1.6.0
termcolor==1.1.0
Werkzeug==1.0.0
